[adapters] Remove non-fault-tolerant Kafka input adapter.
There's no longer any advantage over the fault-tolerant one.
