+++ crates/adapters/src/transport/kafka/nonft/mod.rs
@@ -1,7 +1,3 @@
-mod input;
 mod output;
-#[cfg(test)]
-pub mod test;
 
-pub use input::KafkaInputEndpoint;
 pub use output::KafkaOutputEndpoint;
